fix(*): update course funcs
