+++ src/screens/Dashboard/AssignCourse.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { getDocs, collection } from "firebase/firestore";
 import imgPlaceholder from "./image-placeholder.png";
-import { assignCourse, selectCourse } from "./courseFunctions";
+import { assignCourse, selectCourse } from "../../utils/courseFunctions";
 import { db } from "../../firebase";
 import "./AssignCourse.css";
