+++ src/screens/Dashboard/ViewStudents.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link, useParams, useNavigate } from "react-router-dom";
-import { searchStudent } from "./courseFunctions";
+import { searchStudent } from "../../utils/courseFunctions";
 import SearchBar from "./SearchBar";
 // import Axios from "axios";
 // import "./PendingItems.css";
